requirements.txt
openai>=1.30.0
pydantic>=2.0.0
streamlit>=1.35.0
streamlit-echarts>=0.4.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
rich>=13.0.0
